refactor(+fix): introduce Mapping struct
This change carries cda36dd39 through to the PLIC mode (for the
'c6), with the primary goal being the elimination of the runtime LUT
(especially in the case of the 'c3, where the mapping is trivial).

In doing so, I found it somewhat difficult to keep track of exactly
which `usize` was a priority and which was an interrupt, hence my
attempt here to make use of the existing types as a mnemonic aid.